pyproject.toml
[tool.poetry]
name = "sneaky-bot"
version = "0.1.0"
description = "A robot that saves and forward several kinds of information received in QQ (such as flash images), based on NoneBot2"
authors = ["Well404 <well_404@outlook.com>"]
license = "AGPL-3.0-or-later"
readme = "README.md"
homepage = "https://github.com/Well2333/SneakyBot"
repository = "https://github.com/Well2333/SneakyBot/tree/master/src/sneaky"
packages = [{include = "sneaky_bot"}]

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0b5"
nonebot-adapter-onebot = "^2.1.3"

[tool.poetry.group.dev.dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[tool.nonebot]
plugins = ["sneaky"]

# [[tool.poetry.source]]
# name = "tsinghua"
# url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
# default = true


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
